Move bar hit error meters to their own namespace + Rename BarHitErrorMeter -> SidebarHitErrorMeter
